backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
joblib>=1.4
python-multipart>=0.0.9
pdfplumber>=0.11
pypdf>=4.2
pillow>=10.3
rapidocr-onnxruntime>=1.3
langchain>=0.2
langchain-openai>=0.1
